Add support for creating and deleting a volume
New methods in BrktService:
* getVolume()
* createVolume()
* deleteVolume()

Add support for DELETE to BrktHttpClient and BrktRestClient.
Consolidate the code that reads the response payload into a single
method.  Remove methods that take Class<T> as an argument.  Turns out
that Class is a subclass of Type.

Add new commands to Main for the new volume operations.  Wrap the main
logic in a try/catch clause that handles RuntimeHttpError, so that we
don't log a stack trace when the server returns an error response.

Use Closeables to close the response stream and get rid of Util (for
now).