Add case format conversion library
Add Guava CaseFormat and classes that it depends on.  This will be used
for converting camelCase field names to underscore_notation when sending
requests from the CLI.

Add a dependency on Google's implementation of JSR 305.  This is a small
library that keeps us from having to comment out @Nullable and
@CheckReturnValue in the bits of Guava code that we copy.