Convert camelCase field names to underscore
When processing command line attributes for POST requests, convert
camelCase field names to underscore_notation.  Java convention is
camelCase and our API uses underscores.  This change allows the user to
use either format on the command line.

Rename "elements" to "attrs", since that's a more concise way to
describe the type of data that we're working with.