Add workload request builder
Add RequestBuilder utility class.  This class assembles a map of
attributes for a POST request and verifies that required attributes were
specified.

Add WorkloadRequestBuilder, which provides a type-safe way to build a
request for creating or updating a workload.

Remove field name constants from Workload.  I don't believe that we need
them if we have request builders.