Initial integration test
Add the beginnings of an integration test for the Java SDK.  To run the
integration test, run the following command in the root directory of the
Java SDK:

$ ./run-integration-test --token 60435a8b84104e25bbe13ad7b63f8695 --key
6e08558938f346038c61716ccb98dacf http://10.95.0.1

The run-integration-test script determines the project classpath and
runs the com.brkt.client.IntegrationTest class.  The test creates a
BrktService and uses it to call REST API endpoints using the credentials
and server URI that are specified on the command line.